Create a LoopbackMessagingContext that supports async operation.
Trying to push the setup that AppContext has to a wider range of tests
