+++ src/app/tests/AppTestContext.cpp
@@ -25,10 +25,7 @@ namespace Test {
 
 CHIP_ERROR AppContext::Init()
 {
-    ReturnErrorOnFailure(chip::Platform::MemoryInit());
-    ReturnErrorOnFailure(mIOContext.Init());
-    ReturnErrorOnFailure(mTransportManager.Init("LOOPBACK"));
-    ReturnErrorOnFailure(MessagingContext::Init(&mTransportManager, &mIOContext));
+    ReturnErrorOnFailure(Super::Init());
     ReturnErrorOnFailure(chip::app::InteractionModelEngine::GetInstance()->Init(&GetExchangeManager(), nullptr));
 
     return CHIP_NO_ERROR;
@@ -37,9 +34,7 @@ CHIP_ERROR AppContext::Init()
 CHIP_ERROR AppContext::Shutdown()
 {
     chip::app::InteractionModelEngine::GetInstance()->Shutdown();
-    ReturnErrorOnFailure(MessagingContext::Shutdown());
-    ReturnErrorOnFailure(mIOContext.Shutdown());
-    chip::Platform::MemoryShutdown();
+    ReturnErrorOnFailure(Super::Shutdown());
 
     return CHIP_NO_ERROR;
 }
